06.phrash06/multi_agent_report/pyproject.toml
[project]
name = "multi-agent-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=0.3",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
